+++ resources/rules.php
@@ -2,35 +2,38 @@
 
 declare(strict_types=1);
 
+use Serhii\Ago\Lang;
 use Serhii\Ago\Rule;
 
 /**
- * @return array<string,Rule>
+ * @return Rule[]
  */
-return function (int $num): array {
+return static function (int $num): array {
     $end = $num % 10;
 
     return [
-        'en,nl,de,es,fr' => new Rule(
+        new Rule(
+            langs: [Lang::EN, Lang::NL, Lang::DE],
             zero: $num === 0,
             one: $num === 1,
             two: $num === 2,
             few: $num > 1,
             many: $num > 1,
         ),
-        'ru,uk,be' => new Rule(
+        new Rule(
+            langs: [Lang::RU, Lang::UK],
             zero: $num === 0,
             one: $num === 1 || ($num > 20 && $end === 1),
             two: $num === 2,
             few: ($end === 2 || $end === 3 || $end === 4) && ($num < 10 || $num > 20),
             many: ($num >= 5 && $num <= 20) || $end === 0 || $end >= 5,
         ),
-        'zh,ja' => new Rule(
-            zero: true,
-            one: true,
-            two: true,
-            few: true,
-            many: true,
-        ),
+        // 'zh,ja' => new Rule(
+        //     zero: true,
+        //     one: true,
+        //     two: true,
+        //     few: true,
+        //     many: true,
+        // ),
     ];
 };