+++ src/Lang.php
@@ -4,20 +4,11 @@ declare(strict_types=1);
 
 namespace Serhii\Ago;
 
-final class Lang
+enum Lang: string
 {
-    public const EN = 'en';
-    public const RU = 'ru';
-    public const UK = 'uk';
-    public const NL = 'nl';
-    public const DE = 'de';
-
-    /**
-     * @param array<int,LangOverwrite>|null $overwrites Overwrite default
-     * translations.
-     */
-    public static function set(string $lang, array|null $overwrites = []): void
-    {
-        TimeAgo::configure(new Config(lang: $lang, overwrites: $overwrites));
-    }
+    case EN = 'en';
+    case RU = 'ru';
+    case UK = 'uk';
+    case NL = 'nl';
+    case DE = 'de';
 }