+++ tests/Translations/DutchTest.php
@@ -15,7 +15,7 @@ final class DutchTest extends TestCase
     #[DataProvider('providerForReturnsCorrectTimeFromOneMinuteAndAbove')]
     public function testReturnsCorrectTimeFromOneMinuteAndAbove(string $input, string $expect): void
     {
-        TimeAgo::configure(new Config(lang: Lang::NL));
+        TimeAgo::reconfigure(new Config(lang: Lang::NL));
         $this->assertSame($expect, TimeAgo::trans($input));
     }
 
@@ -63,7 +63,7 @@ final class DutchTest extends TestCase
     #[DataProvider('providerForReturnsCorrectDateFrom0SecondsTo59Seconds')]
     public function testProviderForReturnsCorrectDateFrom0SecondsTo59Seconds(int $seconds, array $expect): void
     {
-        TimeAgo::configure(new Config(lang: Lang::NL));
+        TimeAgo::reconfigure(new Config(lang: Lang::NL));
 
         $result = TimeAgo::trans("now - {$seconds} seconds");
         $this->assertContains($result, $expect);