+++ tests/Translations/RussianTest.php
@@ -15,7 +15,7 @@ final class RussianTest extends TestCase
     #[DataProvider('providerForReturnsCorrectTimeFromOneMinuteAndAbove')]
     public function testReturnsCorrectTimeFromOneMinuteAndAbove(string $input, string $expect): void
     {
-        TimeAgo::configure(new Config(lang: Lang::RU));
+        TimeAgo::reconfigure(new Config(lang: Lang::RU));
         $this->assertSame($expect, TimeAgo::trans($input));
     }
 
@@ -94,7 +94,7 @@ final class RussianTest extends TestCase
     #[DataProvider('providerForReturnsCorrectDateFrom0SecondsTo59Seconds')]
     public function testReturnsCorrectDateFrom0SecondsTo59Seconds(int $seconds, array $expect): void
     {
-        TimeAgo::configure(new Config(lang: Lang::RU));
+        TimeAgo::reconfigure(new Config(lang: Lang::RU));
 
         $result = TimeAgo::trans("now - {$seconds} seconds");
         $this->assertContains($result, $expect);