+++ tests/Translations/UkrainianTest.php
@@ -15,7 +15,7 @@ final class UkrainianTest extends TestCase
     #[DataProvider('providerForReturnsCorrectTimeFromOneMinuteAndAbove')]
     public function testReturnsCorrectTimeFromOneMinuteAndAbove(string $input, string $output): void
     {
-        TimeAgo::configure(new Config(lang: Lang::UK));
+        TimeAgo::reconfigure(new Config(lang: Lang::UK));
         $this->assertSame($output, TimeAgo::trans($input));
     }
 
@@ -84,7 +84,7 @@ final class UkrainianTest extends TestCase
     #[DataProvider('providerForReturnsCorrectDateFrom0SecondsTo59Seconds')]
     public function testReturnsCorrectDateFrom0SecondsTo59Seconds(int $seconds, array $expect): void
     {
-        TimeAgo::configure(new Config(lang: Lang::UK));
+        TimeAgo::reconfigure(new Config(lang: Lang::UK));
 
         $result = TimeAgo::trans("now - {$seconds} seconds");
         $this->assertContains($result, $expect);